母题代码/embedding/requirements.txt
openai>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
plotly>=5.14.0
umap-learn>=0.5.3
jieba>=0.42.1
